Initial commit for DS3231 library.
Lay out API for the library.  Document the registers that we'll be interfacing
with in the library for ease of reference.

Implement the class's first two methods, Read() and SetClock().

Read() reads updated information from the RTC's registers and stores them in
the datastructure.

SetClock() sets the internal clock to an arbitrary time in the current century.